Create and populate a valid EFI partition
This change creates a /boot/efi partition, installs grub2-efi and shim
packages, and generates the grub2 config for both EFI boot and legacy
BIOS boot.

This will allow images to be booted on UEFI boot platforms, even with
Specure Boot enabled.

These changes came from the rhel9 branch of spin-kickstarts before
rhel9/rhel-9.0-kvm-x86_64.ks was deleted. RHEL-9 images are now built
with osbuild-composer[1]. I assume it is possible centos images will
be built with osbuild-composer in the future, so this change also
brings this image incrementally closer to the current partition layout
of rhel-guest-image-9.

[1] https://github.com/osbuild/osbuild-composer/blob/main/internal/distro/rhel90/distro.go